Upgrade to Materialize 0.100.2
This is the last release pre-1.0.0. The v1 release removes JavaScritp
transitions that are currently in use by assets/js/theme.js (copy of
function below).

This commit also updates the mobile navigation code so
that it will continue to work. Doing so included importing the Gooogle
Icon Font. Those font icons and the ones from Font Awesome were used to
replace icons no longer included with Materialize. Font Awesome icons
were chosen for the person, calendar, and forward/back buttons because
they provided better alignment with the text next to them.

Also contained is a fix for alignment in the footer that resulted from
the change to the CSS.

Lastly, the location of the disclaimer text had to be adjusted becuase
the color it inherited changed to white by way of the footer.

```javascript
function ApplyMaterializeAnimation() {
    Materialize.fadeInImage('.content-wrapper');
    Materialize.fadeInImage('.sidebar');
}
```